Move records and table logic to widgets
